requirements.txt
Flask==2.3.3
Pillow==10.0.0
numpy==1.24.3
joblib==1.3.2
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
albumentations==1.3.1
streamlit==1.28.1
